Refac: Adicionado drop-box nas comandas dos clientes: - Espaço otimizado através de ocultamento de informações nas comandas; - Fixado um "Cliente Teste" no HTML; - Pequenas reestilizações nas comandas; - Adicionados novos exports/imports.
