+++ restaurantProject/restaurantStyle.css
@@ -6,8 +6,13 @@
   letter-spacing: 1px;
 }
 
-::selection {
-  background-color: inherit;
+.information:hover {
+  color: darkslategray;
+  opacity: 1;
+}
+
+.information:active {
+  color: darkred;
 }
 
 #reveal-aside-section:hover {
@@ -49,6 +54,12 @@ option {
 .main-section {
   border-bottom: groove thick gray;
   position: relative;
+}
+
+.information {
+  font-size: smaller;
+  opacity: 0.85;
+  cursor: pointer;
 }
 
 #current-clients {
@@ -185,7 +196,9 @@ li.selected {
 }
 
 .client-divs {
+  position: relative;
   font-size: 1.1rem;
+  min-width: 450px;
   max-width: 60vw;
   margin: 25px 5px;
   padding: 15px;
@@ -194,7 +207,31 @@ li.selected {
 }
 
 .client-divs div{
-  margin-top: 10px;
+  margin: 15px 0;
+}
+
+.more-info {
+  padding: 3px;
+  width: 35px;
+  font-size: 1.1rem;
+  font-weight: 600;
+}
+
+.hidden-div {
+  display: none;
+  z-index: 2;
+  position: absolute;
+  left: 0;
+  min-width: 100%;
+  padding: 10px;
+  border: groove lightgrey;
+  outline: groove grey;
+  background-color: rgba(233, 233, 233, 0.95);
+}
+
+.hidden-div div {
+  padding: 5px;
+  
 }
 
 .client-checks {
